+++ app.js
@@ -8,7 +8,6 @@ require('dotenv').config();
 const router = require('./src/routes/index');
 const models = require('./src/db/models/index');
 const pageRouter = require('./src/routes/page');
-const mypageRouter = require("./src/routes/mypage.js")
 
 const env = process.env;
 
@@ -20,7 +19,7 @@ app.set('views', path.join(__dirname, 'src', 'views'));
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 app.use(cookie_parser());
-app.use('/api', [router, mypageRouter]);
+app.use('/api', router);
 app.use('/', pageRouter);
 app.use(express.static(path.join(__dirname, 'src', 'public')));
