+++ package.json
@@ -29,7 +29,6 @@
     "nodemon": "^2.0.20",
     "prettier": "^2.8.3",
     "sequelize-cli": "^6.6.0"
-
   },
   "repository": {
     "type": "git",
@@ -41,7 +40,3 @@
   "homepage": "https://github.com/pmu014/mylovemeat#readme",
   "description": ""
 }
-
-  }
-}
-
