+++ src/routes/products.routers.js
@@ -5,8 +5,8 @@ const ProductsController = require('../controllers/products.controllers');
 const router = express.Router();
 const productsController = new ProductsController();
 
-router.get('/:productId', productsController.getProduct);
+// router.get('/:productId', productsController.getProduct);
 
-router.get('/', productsController.getProducts);
+// router.get('/', productsController.getProducts);
 
 module.exports = router;
